Reorg pagetable rewrite (pt 2).
