+++ src/kernel/src/arch/amd64/memory.rs
@@ -5,6 +5,7 @@ use x86_64::{structures::paging::PageTable};
 use crate::memory::{VirtAddr, PhysAddr};
 
 pub mod frame;
+pub mod pagetables;
 
 fn translate_addr_inner(addr: VirtAddr, phys_mem_offset: VirtAddr) -> Option<PhysAddr> {
     use x86_64::registers::control::Cr3;
