+++ src/kernel/src/arch/amd64/memory/pagetables.rs
@@ -1,0 +1,7 @@
+mod consistency;
+mod entry;
+mod table;
+
+pub use consistency::{ArchCacheLineMgr, ArchTlbMgr};
+pub use entry::{Entry, EntryFlags};
+pub use table::Table;
