+++ src/kernel/src/arch/amd64/memory/pagetables/table.rs
@@ -1,0 +1,78 @@
+use core::ops::{Index, IndexMut};
+
+use crate::arch::address::VirtAddr;
+
+use super::Entry;
+
+#[repr(transparent)]
+pub struct Table {
+    entries: [Entry; Self::PAGE_TABLE_ENTRIES],
+}
+
+impl Table {
+    pub const PAGE_TABLE_ENTRIES: usize = 512;
+    pub fn top_level() -> usize {
+        // TODO: support 5-level paging
+        3
+    }
+
+    pub fn can_map_at_level(level: usize) -> bool {
+        match level {
+            0 => true,
+            1 => true,
+            // TODO: check cpuid
+            2 => true,
+            _ => false,
+        }
+    }
+
+    pub fn set_count(&mut self, count: usize) {
+        // NOTE: this function doesn't need cache line or TLB flushing because the hardware never reads these bits.
+        for b in 0..16 {
+            if count & (1 << b) == 0 {
+                self[b].set_avail_bit(false);
+            } else {
+                self[b].set_avail_bit(true);
+            }
+        }
+    }
+
+    pub fn read_count(&self) -> usize {
+        let mut count = 0;
+        for b in 0..16 {
+            let bit = self[b].get_avail_bit();
+            count |= if bit { 1 } else { 0 } << b;
+        }
+        count
+    }
+
+    pub fn is_leaf(addr: VirtAddr, level: usize) -> bool {
+        level == 0 || addr.is_aligned_to(1 << (12 + 9 * level))
+    }
+
+    pub fn get_index(addr: VirtAddr, level: usize) -> usize {
+        let shift = 12 + 9 * level;
+        (u64::from(addr) >> shift) as usize & 0x1ff
+    }
+
+    pub fn level_to_page_size(level: usize) -> usize {
+        if level > 3 {
+            panic!("invalid level");
+        }
+        1 << (12 + 9 * level)
+    }
+}
+
+impl Index<usize> for Table {
+    type Output = Entry;
+
+    fn index(&self, index: usize) -> &Self::Output {
+        &self.entries[index]
+    }
+}
+
+impl IndexMut<usize> for Table {
+    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
+        &mut self.entries[index]
+    }
+}
